scribe argument

ReferenceClass object for managing arguments

Fields

aliases

[character]
A vector to denote the argument's name

action

[character]
An action for resolving the argument

options

[list]
A named list of options (see Options)

convert

[ANY]
Passed to the to argument in value_convert()

default

[ANY]
A default value

info

[character]
Additional information about the argument when printed

n

[integer]
The length of the values

positional

[logical]
Indicator if the argument is positional (i.e., not preceded by a - or -- command line argument)

Methods

get_action()

Retrieve action

get_aliases()

Retrieve aliases

get_default()

Retrieve the default value

get_help()

Retrieve help information as a character vector

get_name(clean = TRUE)

Retrieve names

clean

When TRUE removes -s from text

help()

Print out formatted help information

initialize( aliases = NULL, action = NULL, options = NULL, convert = NULL, default = NULL, info = NULL, n = NULL )

Initialize the scribeArg object

See fields for parameter information.

parse_value(command_arg)

Parse argument value

This method will likely not be called directly by the user. Instead, this method is called within scribeCommandArgs.

command_arg

A scribeCommandArgs

object

Options

Several available options

action="list"

choices

An explicit set of values that argument must be. If the value parsed is not one of these, an error will occur.

action="flag"

no

When TRUE included appends --no to aliases to invert results
Example:
With the argument new_arg("--test", options = list(no = TRUE)), passing command arguments --test would set this to TRUE and --no-test explicitly set to FALSE.